% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_coords}
\alias{city_coords}
\title{Coodinates of European Cities}
\format{A data frame with 379 rows and 3 columns
\describe{
 \item{\code{lat}}{Double. Latitude.}
 \item{\code{lon}}{Double. Longitude.}
 \item{\code{city}}{Character. City name.}
}}
\usage{
city_coords
}
\description{
Coodinates of European Cities
}
\note{
This table can be joined to the \link{europop} table by the \code{city} column.
}
\keyword{datasets}

